added framework for compile_pred PTestInstr and PTestRes.
I'll be completely honest, it's very hard to understand what I'm doing when I cant do a test.  I managed to define a PZero and a Buffer but compile_pred says that the printing of that is not possible, for some reason.  Due to (Show (CompileM Exp)).  deriving Show doesn't seem to help anywhere.