Added compile_pred for Neg and Prod
Known errors:
Neg doesn't actually test value A (pred1) because that might be EITE but I don't remember what EITE does
Sort of the same thing with EAnd in Prod.  It should be comparing 2 bools returned by two Exp?  One of the Exp functions should return a bool I think.  Or something related to the predicates
	It might be part of the predicate definition or the testing of values.